policy commit

+++ app/Http/Controllers/PrescriptionController.php
@@ -28,20 +28,24 @@ class PrescriptionController extends Controller
 
     public function getPrescriptionCreatedByDoctor($id)
     {
+       
         $prescriptions = $this->prescriptionService->getPrescriptionListByDoctorId($id);
+        $this->authorize('viewAny', Prescription::class); 
         return view('prescription.index',compact('prescriptions'));
     }
 
     public function getPrescriptionByPatientId($id)
     {
-
+       
         $prescriptions = $this->prescriptionService->getPrescriptionListByPatientId($id);
+        $this->authorize('viewAny', Prescription::class); 
         return view('prescription.index',compact('prescriptions'));
     }
 
     
     public function create()
     {
+        $this->authorize('create', Prescription::class); 
         $doctor= $this->doctorService->getDoctorByUserId(Auth()->user()->id);
         $appointments = $this->appointmentService->getAppointedPatientDropdownByDoctorId($doctor->id);
         return view('prescription.create',compact('appointments','doctor'));
@@ -49,8 +53,8 @@ class PrescriptionController extends Controller
 
     public function store(PrescriptionRequest $request)
     {
-        $validatedData = $request->validated();
-        
+        $this->authorize('create', Prescription::class); 
+        $validatedData = $request->validated();   
         $this->prescriptionService->createOrUpdate($validatedData);
         Session::flash('message','Information Added successfully!!!!');
         return redirect()->route('prescription.doctor',$validatedData['doctor_id']);
@@ -59,6 +63,7 @@ class PrescriptionController extends Controller
     public function edit($id)
     {
         $prescription = $this->prescriptionService->getById($id);
+        $this->authorize('update',$prescription); 
         $doctor= $this->doctorService->getDoctorByUserId(Auth()->user()->id);
         $appointments = $this->appointmentService->getAppointedPatientDropdownByDoctorId($doctor->id);          
         return view('prescription.edit', compact('prescription','doctor','appointments'));
@@ -66,6 +71,8 @@ class PrescriptionController extends Controller
     
     public function update(PrescriptionRequest $request, $id)
     {
+        $prescription = $this->prescriptionService->getById($id);
+        $this->authorize('update',$prescription);
         $validatedData = $request->validated();
         $validatedData['id']=$id;//precsription id
         
@@ -78,6 +85,7 @@ class PrescriptionController extends Controller
     public function show($id){
 
         $prescription = $this->prescriptionService->getById($id);
+        $this->authorize('view',$prescription);
         return view('prescription.show',compact('prescription'));
     }
 
@@ -85,6 +93,8 @@ class PrescriptionController extends Controller
     public function destroy($id)
     {
 
+        $prescription = $this->prescriptionService->getById($id);
+        $this->authorize('delete',$prescription);
         $doctor= $this->doctorService->getDoctorByUserId(Auth()->user()->id);
         $prescription = $this->prescriptionService->delete($id);
